updated documentation

+++ src/boardgenerator/GraphUtil.java
@@ -13,9 +13,10 @@ public class GraphUtil {
      * @param d1       the first double to compare
      * @param d2       the second double to compare
      * @param accuracy the accuracy to compare with
-     * @return the result of the comparison
+     * @return After adjusting values based on accuracy, returns {@code 0} if values are equal,
+     * negative integer if d1 is less than d2, positive if d1 is greater than d2
      */
     public static int compare(double d1, double d2, double accuracy) {
-        return Float.compare(round(d1 / accuracy), round(d2 / accuracy));
+        return Double.compare(round(d1 / accuracy), round(d2 / accuracy));
     }
 }
